Adding support for filters.
Now customizations can implement filtering functions, and
register them, which in turn the core code calls. The core
does not have to be altered when a new filtering function
is added. New types of filters, however, have to be added to core
for them to be called.

Use with care :-)